DANGER: changing co2 and delay types to int. Using migraiton scripts for database

+++ src/main/java/com/breathe/model/ApiDeviceModel.java
@@ -4,13 +4,13 @@ package com.breathe.model;
  * Created by denis on 07.04.15.
  */
 public class ApiDeviceModel extends DeviceModel {
-    public double currentCO2;
+    public int currentCO2;
 
     public double currentTemperature;
 
     public double currentHumidity;
 
-    public double deltaCO2;
+    public int deltaCO2;
 
     public double deltaTemperature;
 
@@ -24,11 +24,11 @@ public class ApiDeviceModel extends DeviceModel {
         this.currentHumidity = currentHumidity;
     }
 
-    public double getCurrentCO2() {
+    public int getCurrentCO2() {
         return currentCO2;
     }
 
-    public void setCurrentCO2(double currentCO2) {
+    public void setCurrentCO2(int currentCO2) {
         this.currentCO2 = currentCO2;
     }
 
@@ -40,11 +40,11 @@ public class ApiDeviceModel extends DeviceModel {
         this.currentTemperature = currentTemperature;
     }
 
-    public double getDeltaCO2() {
+    public int getDeltaCO2() {
         return deltaCO2;
     }
 
-    public void setDeltaCO2(double deltaCO2) {
+    public void setDeltaCO2(int deltaCO2) {
         this.deltaCO2 = deltaCO2;
     }
